added check for hosts file

+++ setup.sh
@@ -12,24 +12,33 @@ NC='\033[0m' # No Color
 # Move to top level of repo
 cd "$(git rev-parse --show-toplevel)"
 
-# Get the pi / vm ip address from user
-echo "What is the IP address for your Raspberry Pi / VM you wish to setup?"
-read -r ipaddr
-
 # Create a hosts file from default
 hostsfile="ansible/playbooks/hosts"
-rm -f $hostsfile
-cp -rp ansible/playbooks/hosts.default $hostsfile
 
-
-# Add the IP address into hosts file
-if [[ $ipaddr != "" ]]; then 
-sed -i '' "s/IPADDRESS/$ipaddr/" $hostsfile 
-else
-echo -e "${RED}ERROR:${NC} user input for IP Address was blank. Please rerun and enter IP Address."
-exit 1
+# Get the pi / vm ip address from user
+if [[ -f $hostsfile ]]; then
+    echo -e "${YELLOW}$hostsfile${NC} already exists, override and create a new hosts file? [y/n]"
+    read -r yn
+    case $yn in
+        [Yy]* )
+                # Replace current hostfile with the default
+                rm -f $hostsfile
+                cp -rp ansible/playbooks/hosts.default $hostsfile
+                # Get users raspberry pi / VM IP address
+                echo "What is the IP address for your Raspberry Pi / VM you wish to setup?"
+                read -r ipaddr
+                # Add the IP address into hosts file
+                if [[ $ipaddr != "" ]]; then 
+                    sed -i '' "s/IPADDRESS/$ipaddr/" $hostsfile 
+                else
+                    echo -e "${RED}ERROR:${NC} user input for IP Address was blank. Please rerun and enter IP Address."
+                    exit 1
+                fi
+                echo "User inputtted IP Address added to $hostsfile.";;
+        [Nn]* ) ;;
+        * ) echo "Please answer yes or no."; exit 1;;
+    esac
 fi
-echo "User inputtted IP Address added to $hostsfile."
 
 # Create local variables for ansible installation
 echo "Using ansible to setup Raspberry Pi / VM."
